NCL-2049: Added websocket endpoints and basic tests for reports/align and reports/built
