--- bc-rest/src/main/java/org/jboss/da/bc/ws/Method.java
@@ -1,14 +1,0 @@
-package org.jboss.da.bc.ws;
-
-/**
- *
- * @author Honza Brázdil <jbrazdil@redhat.com>
- */
-public interface Method<T, S> {
-
-    public String getName();
-
-    public Class<T> getParameterClass();
-
-    public S execute(T params) throws Exception;
-}
