+++ testsuite/src/test/java/org/jboss/da/test/client/ws/reports/AbstractWebsocketReportsTest.java
@@ -1,0 +1,33 @@
+package org.jboss.da.test.client.ws.reports;
+
+import static org.jboss.da.common.Constants.REST_API_VERSION_BC;
+
+import org.jboss.arquillian.container.test.api.Deployment;
+import org.jboss.da.test.ArquillianDeploymentFactory;
+import org.jboss.da.test.ArquillianDeploymentFactory.DepType;
+import org.jboss.da.test.client.ws.AbstractWebsocketApiTest;
+import org.jboss.shrinkwrap.api.spec.EnterpriseArchive;
+
+public abstract class AbstractWebsocketReportsTest extends AbstractWebsocketApiTest {
+
+    private static final String DEFAULT_REST_API_VERSION = "v-" + REST_API_VERSION_BC;
+
+    @Deployment
+    public static EnterpriseArchive createDeployment() {
+        return new ArquillianDeploymentFactory().createDeployment(DepType.REPORTS);
+    }
+
+    @Override
+    protected String getContextRoot() {
+        return "testsuite";
+    }
+
+    protected String readRestApiVersion() {
+        return readConfigurationValue("testsuite.restApiVersion", DEFAULT_REST_API_VERSION);
+    }
+
+    @Override
+    protected String convertRestApiVersionToFolderName() {
+        return getFolderName(DEFAULT_REST_API_VERSION);
+    }
+}
