+++ testsuite/src/test/rest/v-1.0/expectedResponse/reports/built/dependency-analysis.json
@@ -1,0 +1,220 @@
+[
+    {
+        "groupId": "com.fasterxml.jackson.core",
+        "artifactId": "jackson-databind",
+        "version": "2.4.4",
+        "builtVersion": null,
+        "availableVersions": [
+            "2.5.4.redhat-1"
+        ]
+    },
+    {
+        "groupId": "org.jboss.spec",
+        "artifactId": "jboss-javaee-6.0",
+        "version": "3.0.2.Final",
+        "builtVersion": "3.0.2.Final-redhat-15",
+        "availableVersions": [
+            "3.0.0.Final-redhat-1",
+            "3.0.2.Final-redhat-7",
+            "3.0.2.Final-redhat-11",
+            "3.0.1.Final-redhat-1",
+            "3.0.2.Final-redhat-3",
+            "3.0.2.Final-redhat-4",
+            "3.0.2.Final-redhat-13",
+            "3.0.2.Final-redhat-15"
+        ]
+    },
+    {
+        "groupId": "org.jboss.da",
+        "artifactId": "bc-backend",
+        "version": "0.1.0",
+        "builtVersion": null,
+        "availableVersions": []
+    },
+    {
+        "groupId": "org.jboss.arquillian.protocol",
+        "artifactId": "arquillian-protocol-servlet",
+        "version": "1.1.8.Final",
+        "builtVersion": null,
+        "availableVersions": []
+    },
+    {
+        "groupId": "org.jboss.shrinkwrap.resolver",
+        "artifactId": "shrinkwrap-resolver-impl-maven-archive",
+        "version": "2.1.1",
+        "builtVersion": null,
+        "availableVersions": []
+    },
+    {
+        "groupId": "org.jboss.da",
+        "artifactId": "common",
+        "version": "0.1.0",
+        "builtVersion": null,
+        "availableVersions": []
+    },
+    {
+        "groupId": "org.jboss.as",
+        "artifactId": "jboss-as-arquillian-container-managed",
+        "version": "7.2.0.Final",
+        "builtVersion": null,
+        "availableVersions": [
+            "7.5.2.Final-redhat-2",
+            "7.5.3.Final-redhat-2",
+            "7.5.7.Final-redhat-3",
+            "7.5.0.Final-redhat-21",
+            "7.5.5.Final-redhat-3",
+            "7.5.6.Final-redhat-2",
+            "7.2.3.Final-redhat-2",
+            "7.2.2.Final-redhat-4",
+            "7.2.2.Final-redhat-1",
+            "7.5.8.Final-redhat-2"
+        ]
+    },
+    {
+        "groupId": "org.slf4j",
+        "artifactId": "slf4j-api",
+        "version": "1.7.7",
+        "builtVersion": "1.7.7.redhat-3",
+        "availableVersions": [
+            "1.6.1-redhat-1",
+            "1.7.2.redhat-2",
+            "1.7.2-redhat-1",
+            "1.7.7.redhat-3",
+            "1.7.2.redhat-3",
+            "1.6.1-redhat-2"
+        ]
+    },
+    {
+        "groupId": "org.jboss.da",
+        "artifactId": "reports-backend",
+        "version": "0.1.0",
+        "builtVersion": null,
+        "availableVersions": []
+    },
+    {
+        "groupId": "org.hibernate",
+        "artifactId": "hibernate-core",
+        "version": "4.2.14.SP4-redhat-1",
+        "builtVersion": "4.2.14.SP4-redhat-1",
+        "availableVersions": [
+            "5.0.9.Final-redhat-1",
+            "4.1.3.Final-redhat-1",
+            "4.2.7.SP1-redhat-3",
+            "4.2.7.SP5-redhat-1",
+            "4.2.14.SP1-redhat-1",
+            "4.2.21.Final-redhat-1",
+            "4.2.0.Final-redhat-1",
+            "4.2.23.Final-redhat-1",
+            "4.2.19.Final-redhat-1",
+            "4.2.17.SP1-redhat-1",
+            "4.2.14.SP4-redhat-1",
+            "4.2.18.Final-redhat-2",
+            "4.2.18.Final-redhat-1",
+            "4.2.14.SP3-redhat-1",
+            "4.2.0.SP1-redhat-1",
+            "4.2.22.Final-redhat-1",
+            "4.1.6.Final-redhat-3",
+            "4.2.20.Final-redhat-1",
+            "4.2.7.SP4-redhat-1"
+        ]
+    },
+    {
+        "groupId": "org.projectlombok",
+        "artifactId": "lombok",
+        "version": "1.16.4",
+        "builtVersion": null,
+        "availableVersions": []
+    },
+    {
+        "groupId": "org.jboss.da",
+        "artifactId": "bc-rest",
+        "version": "0.1.0",
+        "builtVersion": null,
+        "availableVersions": []
+    },
+    {
+        "groupId": "org.jboss.da",
+        "artifactId": "communication",
+        "version": "0.1.0",
+        "builtVersion": null,
+        "availableVersions": []
+    },
+    {
+        "groupId": "org.jboss.arquillian.junit",
+        "artifactId": "arquillian-junit-container",
+        "version": "1.1.8.Final",
+        "builtVersion": null,
+        "availableVersions": []
+    },
+    {
+        "groupId": "org.jboss.resteasy",
+        "artifactId": "resteasy-client",
+        "version": "3.0.11.Final",
+        "builtVersion": null,
+        "availableVersions": [
+            "3.0.16.Final-redhat-1"
+        ]
+    },
+    {
+        "groupId": "javax.enterprise",
+        "artifactId": "cdi-api",
+        "version": "1.2",
+        "builtVersion": "1.2.0.redhat-2",
+        "availableVersions": [
+            "1.0-SP4-redhat-4",
+            "1.0-SP4.redhat-3",
+            "1.0.0.SP4-redhat-5",
+            "1.0.0.SP4-redhat-6",
+            "1.2.0.redhat-2",
+            "1.0-SP4-redhat-1",
+            "1.0-SP4-redhat-2"
+        ]
+    },
+    {
+        "groupId": "org.jboss.resteasy",
+        "artifactId": "resteasy-jaxrs",
+        "version": "3.0.11.Final",
+        "builtVersion": null,
+        "availableVersions": [
+            "3.0.16.Final-redhat-1",
+            "2.3.3.Final-redhat-1",
+            "2.3.8.SP3-redhat-2",
+            "2.3.7.1.Final-redhat-1",
+            "2.3.7.2.Final-redhat-1",
+            "2.3.7.Final-redhat-2",
+            "2.3.8.SP4-redhat-2",
+            "2.3.10.Final-redhat-1",
+            "2.3.12.Final-redhat-1",
+            "2.3.4.Final-redhat-2",
+            "2.3.8.SP2-redhat-3",
+            "2.3.8.SP2-redhat-1",
+            "2.3.8.Final-redhat-3",
+            "2.3.13.Final-redhat-1",
+            "2.3.6.Final-redhat-1"
+        ]
+    },
+    {
+        "groupId": "org.jboss.da",
+        "artifactId": "reports-rest",
+        "version": "0.1.0",
+        "builtVersion": null,
+        "availableVersions": []
+    },
+    {
+        "groupId": "javax.ws.rs",
+        "artifactId": "javax.ws.rs-api",
+        "version": "2.0",
+        "builtVersion": null,
+        "availableVersions": []
+    },
+    {
+        "groupId": "junit",
+        "artifactId": "junit",
+        "version": "4.11",
+        "builtVersion": "4.11-redhat-1",
+        "availableVersions": [
+            "4.10-redhat-2",
+            "4.11-redhat-1"
+        ]
+    }
+]
