+++ testsuite/src/test/rest/v-1.0/request/reports/built/dependency-analysis.json
@@ -1,0 +1,6 @@
+{
+   "scmUrl": "https://github.com/project-ncl/dependency-analysis.git",
+   "revision": "05a878",
+   "additionalRepos": [],
+   "pomPath": ""
+}
